requirements.txt
boto3==1.34.93
chalice==1.31.0
pyrebase4==4.7.1
